bw/requirements.txt
appnope==0.1.0
backports.shutil-get-terminal-size==1.0.0
baidu-aip==2.0.0.1
beautifulsoup4==4.6.0
bs4==0.0.1
certifi==2017.11.5
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
decorator==4.1.2
Django==1.11.3
django-simple-serializer==2.0.7
djangorestframework==3.7.7
enum34==1.1.6
future==0.16.0
html5lib==1.0.1
idna==2.6
ipython==5.3.0
ipython-genutils==0.2.0
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.0
pathlib2==2.3.0
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.15
psycopg2==2.7.1
ptyprocess==0.5.2
Pygments==2.2.0
pytz==2017.2
requests==2.18.4
scandir==1.5
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.2
uritemplate==3.0.0
urllib3==1.22
wcwidth==0.1.7
webencodings==0.5.1
